finance-chat-bot/requirements.txt
langchain
pinecone-client
python-dotenv
streamlit
pypdf
sentence-transformers